requirements.txt
requests
configparser
google-cloud-language
google-cloud-translate
tweepy
textblob
boto3
